vim-patch:7.4.845
Problem:    Compiler warning for possible loss of data.
Solution:   Add a type cast. (Erich Ritz)

https://github.com/vim/vim/commit/5df1ed2de3fa9dcace996b9a0a4c9b3cea79cf1e

+++ src/nvim/os/env.c
@@ -278,7 +278,7 @@ void expand_env_esc(char_u *srcp, char_u *dst, int dstlen, bool esc, bool one,
       }
       memcpy((char *)dst, (char *)var, len);
       dst += len;
-      dstlen -= len;
+      dstlen -= (int)len;
       continue;
     }
     copy_char = true;
